requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backports.functools-lru-cache==1.6.4
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cn2an==0.5.19
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
demucs==4.0.0
diffq==0.2.3
dora-search==0.1.11
einops==0.6.0
eng-to-ipa==0.0.2
entrypoints==0.4
fastapi==0.95.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.10.7
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
google-auth==2.17.0
google-auth-oauthlib==0.4.6
gradio==3.23.0
grpcio==1.53.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
indic-transliteration==2.3.37
inflect==6.0.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
julius==0.2.7
kiwisolver==1.4.4
ko-pron==1.3
lameenc==1.4.2
librosa==0.9.1
linkify-it-py==2.0.0
llvmlite==0.39.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==9.1.0
mp3play==0.1.15
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
num-thai==0.0.5
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
openai == 0.27.2
openai-whisper==20230314
OpenCC==1.1.1
openunmix==1.2.1
orjson==3.8.9
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.2.0
playsound==1.3.0
pooch==1.7.0
proces==0.1.4
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
pyopenjtalk==0.3.0
pyparsing==3.0.9
pypinyin==0.48.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
retrying==1.3.4
rfc3986==1.5.0
roman==4.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
starlette==0.26.1
submitit==1.4.5
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tiktoken==0.3.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
tqdm==4.65.0
treetable==0.2.5
typer==0.7.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
Unidecode==1.3.6
urllib3==1.26.15
uvicorn==0.21.1
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
